Support for `--live` or `--file` flags
The server can be now started with either a `--live` flag or a `--file`
flag. The live flag can be used to get the data from the interface to
the browser via `redis queue`. A lot of hard-coded part is now removed.

So the server runs in live capture mode or the pcap mode. Also added a
new small pcap file with 20 packets for testing, as the packet format in
the original file `small.json` is not compliant with the supported one.